Decommission the EditorActivation instrument
The instrument was created ~October 2015. Since then, however, the data
that the instrument collected has been replicated in the
wmf.active_editors table [0] and all reports about active editors by WMF
are generated from the data in that table.

Further to the above, there are no mentions of EditorActivation outside
of the instrument's definition, the sanitization allowlist, and the
SAL [1][2][3].

[0] https://github.com/wikimedia/analytics-wmf-product-jobs/blob/master/movement_metrics/notebooks/supplementary/queries/update_active_editors.hql
[1] https://codesearch.wmcloud.org/search/?q=EditorActivation
[2] https://global-search.toolforge.org/?q=EditorActivation
[3] https://github.com/search?q=org%3Awikimedia-research+EditorActivation&type=code

Bug: T330766
Depends-On: If5c0d701d58ade81cdcd6cc79e3f35d2e4780d35
Change-Id: I218d02076e39d300c0c3591c1f69c4cde63eeddf