desktopWebUIActions: Remove disabled abstraction for getEnabledClass
This is a left-over from a previous iteration of I9e8b8ae43d382436a8e
where the function also had a "disabled" version. This no longer exists
but the parameter and function still existed, thus leaving behind a way
to use it incorrectly. E.g. passing `false` or forgetting to pass true
(undefined=falsey) would result in an incomplete class name, ending
with a '-' dash.

While at it, inline it at the single place that uses it, but leave a
separate variable for readability.

Follows-up I9e8b8ae43d382436a8e.

Change-Id: I80fd05e93a512a8bf6935d75fbfc7674c429416e