Revert "*WebUIActionsTracking: Also log events via Metrics Platform"
This reverts commit bdbef33042280a2b734a53f81832585b005cb10c.

Reason for revert: Removing the partial migration of the instrument to minimise our own maintenance burden. It uses dispatch() which is now deprecated in favour of submitInteraction(). We (Data Products) are currently discussing collaborating with the owners of the instrument, Web, who have signalled that they want it to be stable while they establish new baseline metrics for an upcoming project

Bug: T351195
Change-Id: Ic7867cb312dcb259a4b14e6671d5323520e92084