Remove partial migration of VisualEditorFeatureUse instrument
There is a partial migration of the VisualEditorFeatureUse instrument to
the MP client library. It uses dispatch(), which is now deprecated in
favour of submitInteraction(). We should remove the partial migration of
the instrument to minimise our own maintenance burden.

In this specific change we are removing the code of the instrument where
we were using the already deprecated dispatch() method.

Bug: T351337
Change-Id: I6af303c26b505f434048b0a4462c6f214bcb4dfb