Track impressions, success and failure on the Special:CreateAccount and Special:Login pages.
This change adds analytics for temporary users
who try to create an account or log in from the
Special:CreateAccount and Special:Login pages.
It relies on an an event schema that includes
the following major fields:

- event_type: either 'impression', 'success' or 'failure'
- page_title: either 'Special:CreateAccount' or 'Special:Login'
- is_temp: whether the user is a temporary user or not
- error_message_key: the reason for the failure, if any

among other basic fields allowed by the schema.
The events are used to measure the conversion rate
of temporary users to registered users, and to identify
any issues or barriers in the account creation or login process.

Bug: T346327
Depends-On: I55dd09a5d2a839464e49466da94403d935d665e2
Change-Id: I51be116eab6a49c968a44e03a54e40d2a2f9550a