authevents: Remove 'centrallogin' entry point
This never really made any sense. Per T91701#1462705 the idea was
to get a sense of whether people need to perform duplicate logins
because of some CentralAuth failure, but it doesn't work at all
in practice - other than stewards, no one visits loginwiki so all
it picked up was bots.

(This only affects the entrypoint part of the statsd key, not the
event part, where the value 'centralauth' is still used for
dedicated metrics like edge login.)

Change-Id: Id51e4f642fa05dad76969b905820484b87b6dec9