Add revision ID to X-Analytics header
[EXT] Extend onXAnalyticsSetHeader to include revision ID

This change extends the onXAnalyticsSetHeader hook in the
WikimediaEvents extension to include the latest revision ID
of the current page in the X-Analytics header. This new detail
will be used to analyze pageviews for content awaiting reversion,
thus aiding the Moderator Tools team in evaluating the potential
impact of Automoderator's swift revert actions on reader exposure
to vandalism.

Bug: T346350
Change-Id: I2b4e6567a24684dc35048f0be20f82336e5b94f1