modules: Disable camelcase errors for properties
The Event Platform (and the Metrics Platform by extension) require event
property names to be snake case [0]. This has led to instrument
maintainers repeatedly disabling/enabling the camelcase rule within
their instruments, which is noisy and error prone.

Disable the camelcase rule for properties and remove the

  /* eslint-{enable,disable} camelcase */

comments wherever possible.

[0]: https://wikitech.wikimedia.org/wiki/Event_Platform/Schemas/Guidelines#No_Capital_letters_-_Use_snake_case

Bug: T347357
Change-Id: I88be8b2a71601a84234927c3e6003effaf6400ef