feat(clientError.js): allow adding custom context to logged errors
When logging application errors, it is often useful to record additional
structured information, specific to the context in which the error
occurred.

Logstash supports this by allowing for "normalized messages" that
reference the additional context instead of including it. This is
already widely used in the server-side error recording, and this change
aims to bring this to the client side as well. The goal is for the
following code to do the intuitively expected thing:

```
const error = new Error('Unexpected error in component {error_context.vueComponentName} in lifecycle stage "{error_context.vueLifecycleInfo}" with message "{error_context.originalErrorMessage}"');
error.error_context = {
  vueComponentName: "MyComponent",
  vueLifecycleInfo: "setup function",
  originalErrorMessage: "undefined is not a function"
};
mw.errorLogger.logError( error, 'error.myextension' );
```

This code now makes use of data in the `error_context` field of the
error object, if it exists. This is a non-standard field name and so is
unlikely to be already in use by existing code. Thus this change should
defacto backwards compatible.

The change to tsconfig is necessary to make it understand Object.assign.
MediaWiki core already has a hard requirement of a superset of ES2015,
so raising it to that level should not be a problem.

Change-Id: I0226f664af8fa223e3d151b7e1fe8873a4781ae7