Restore cross-wiki upload messages
Partially reverts Iba050434df7f6768b52c386185cf9f632fd6d8a5.

The messages for cross-wiki upload were previously thought to be
unused, but they are used.

The tag is assigned in onRecentChangeSaveCrossWikiUpload() in
WikimediaEventsHooks.php, and the message is used indirectly in
tagShortDescriptionMessage() and elsewhere in
includes/changetags/ChangeTags.php of MediaWiki.

Change-Id: Ice18f8f731bb8adb09b539ba2ab7942223a6aea0