search satisfaction: track virtual page views
We want to track virtual page views as a potential source of
search satisfaction. To support this add a new action type to the
search satisfaction schema which subscribes to the existing
virtual page view event and  produces it to search satisfaction
along with the general metadata we typically produce.

This was done as a subscribe and reproduce, rather than backend
joining of the data, as we need the context of the pageViewId and
searchSessionId to integrate the virtual page view into the
existing metrics pipelines.

Bug: T375387
Depends-On: Ib7b57be5538b6cae102662ccabedf8f54fd9e783
Change-Id: I15979c066f99630b9bc08e6e0c53a58978a9fa89