Experimental: Introduce sessionLength mixin
Adds a sessionLength instrument, which is largely similar
to sessionTick.js, with the primary difference being
that it allows for session length data to be sent to a
specific, user-defined Metric Platform data stream, instead of the
separate `wmf.session_length_daily` table.

This functionality allows feature teams track session length
in conjunction with other data in the same stream, which
enables the use of session length as a guard rail metric
during experimental analysis.

Bug: T378109
Bug: T378072
Bug: T383422
Change-Id: Iabc00220e262c8d47765f2da59320ef7c10cad6c