statsd: Remove subscriber for `mw.track('gauge')`
Introduced in 2015 for use by the WebPageTest application (T112713).
Note that this ran at AWS as a standalone appplication. It was never
used by MediaWiki JS via `mw.track()`, but added for completion.

Shortly after this, we transitioned WebPageTest to have its own
Graphite and Statsd instance locally (with statsvr.py for
compatibility), and shortly after that, transitioned the test reporter
to speak to the local StatsD UDP port directly. See also I7a8d831817,
which removed statsvr.py for the same reason.

Ref https://codesearch.wmcloud.org/search/?q=track%5C%28.%2Agauge

Change-Id: I97321ec537e4dec642b87aba0a2a84c6c6814401