New Web A/B Test Interaction function
Add a new metrics instrumentation file, which will log interactions to
the newly-defined `product_metrics.web_base.search_ab_test_clicks`
stream.  Right now we're using version 1.3.0 of the schema since it
seems to work for validation, but this can always be changed later

For now this implementation is a bit light, so definitely let me know if
there's anything I missed in this ticket!

Additional change:
* Fix unnecessary if statement in index.js relating to search satisfaction
schema

Bug: T383611
Change-Id: Ibdb5af2f2d2dae819794620a5897e79abd5b1dad