search: Augment typeahead url with test parameters
As part of AB testing changes to search we need the ability to
augment the URL being used for typeahead to apply a test
treatment.  In the depended-on patch a hook was added for this
use case, and here we add a handler for said hook.

Note that this is still only best effort, extensions have the
ability to replace the typeahead implementation which would
skip all of this code. The only current implementation of that I
could find is in wikibase, meaning this will not work on
wikis that allow wikibase to replace the typeahead (wikidata.org).

Bug: T397732
Depends-On: I1a477dfaa2b41846cb0167d6639e14b2abbe1e33
Change-Id: Ic0bfb13e8a5ee828c381f0bc6510593b8e3e12ca