search: Ensure msToDisplayResults is an integer
The schema used here requires an integer, but the recent change
in I61a18a85cc to the new timing api's has led to some events
being submitted with non-integer values.

The domInteractive property is a DOMHighResTimeStamp which has a
type of double, use Math.floor to reduce precision to the
expected integer range.

Bug: T400834
Change-Id: I7093adcfe040eb60472200cf595dff63c2baa05f