ClickThroughRateInstrument: introduce flag to prevent interaction bubbling
Prior to the change starting a CTR instrument on two elements with a
parent-child relation will log a click event for both elements when
the events target is contained by both. While this can make sense for
some cases it also difficults the computation of separated CTRs for
this kind of element relation.

Introduce an option trackSingleClick to log a sole click
event attributed to the first registered instrument that contains the
event target and sets the flag to true.

Bug: T403280
Change-Id: I2591db38bd319a11f6b93afdbd02afd45d0ea1c6