statsd: Improve error logs by adding invalid value
This intended to improve the unfortunate situation that we see
errors in logstash about invalid timing values, but not what those
values actually are. We are then left guessing what might be going wrong
for there to be an invalid value which could be any number of things.

This builds on top of the work in I0226f664af8fa223e3d151b7e1fe8873a4781ae7

Bug: T395627
Change-Id: I2901c71f19f3da06de273907e9ea8177de135a02