Port Java Pageview definition to bot detection
The simple-bot-detection instrument's config in xLab has been updated to
include user agent.  Whether or not this is a pageview is being sent via
`action_context` as interaction data.  And the client_ip is being sent by
default with the custom web_base_with_ip schema that the instrument is
now using.

The original pageview logic is defined in:

https://gerrit.wikimedia.org/r/plugins/gitiles/analytics/refinery/source/+/refs/heads/master/refinery-core/src/main/java/org/wikimedia/analytics/refinery/core/PageviewDefinition.java#242

We are implementing it here as closely as possible with what's available
to us in a web context.  This implementation was converted to Python and
run against webrequest data from 2025-09-26 T15.  The return matched
is_pageview 99.6% of the time.  The main mismatch happened when pages
did not set wgRelevantArticleId, for example:

https://api.wikimedia.org/wiki/File:Cooperative_economic_insect_report_(IA_cooperativeecono1822unit).pdf

This makes sense as that page is shadowing the page from Commons and so
does not have a page_id on the local wiki.  But it raises the issue that
these are marked as pageviews in the webrequest dataset - we will circle
back on that.

Bug: T406359
Change-Id: Ib472f45c0c795ac6f4034e1f050e683eb848e983
(cherry picked from commit f4d0665a0a9cd73c6908ebcfe141477f4e8bd7c5)