Explicitly mock User::getWikiId in PHPUnit tests
The method does not have a return type declaration, so PHPUnit will
simply use `null` as the mock return value. But null is not a valid wiki
ID, and it will end up being passed to methods that can't handle it and
ultimately fail as seen in the task. So, mock the value explicitly.

Bug: T409088
Change-Id: Icd81125d8db09841c98d00ec8ae5e2eda0df0fc3