Support Instrument, Experiment in SessionLengthInstrumentMixin
Tweak to the SessionLengthInstrumentMixin to support
sending events to an xLab Experiment or Instrument instead of
via a raw event submission with a stream/schema.

Notes:

* the start() and stop() methods can optionally take an Instrument
  or Experiment in place of a stream string. This object will be used
  as the key in the state map so you pass it back to stop() to end.

* if an object is passed to start(), the second parameter is checked
  for the data object to extend from on data pings.

* the object's `submitInteraction` method is called with the data
  parameter.

* expose the instrument in the `mw.wikimediaEvents` namespace

Note that calls to `submitInteraction` on an Experiment will wrap
the `send` call that we want, so there's no need to distinguish
between the two types of objects here. They are API-compatible.

Bug: T410533
Change-Id: I08b9fb0c9d87d435d9fe74b94daa88a714c4ef52