Refactor: Move editing session ID logic into service
Pull the editing session ID generation out of editAttemptStep.js
into a new editingSessionService.js module to cut duplication and
make the code easier to maintain.

The service exposes a getEditingSessionId() helper that picks the ID
using this order of precedence:

1. mw.config.get('wgWMESchemaEditAttemptStepSessionId')
2. URL parameter 'editingStatsId'
3. Form field '#editingStatsId'
4. Previously passed-in session ID
5. A freshly generated random session ID

This makes it simple to share consistent editing session IDs with
other features, such as hCaptcha instrumentation.

Bug: T406865
Change-Id: If4402199b37bee3153f8a8bbb91d22b584ca5289