feat(mw.track): add support for histogram metrics
The `_distribution` suffix is a generic choice because these histograms
can be about anything, and Wikipedia calls histograms a distribution.

The overall structure is based on Ia693687d1 in the GrowthExperiments
extension (commit f85c832), \Wikimedia\Stats\Metrics\HistogramMetric,
and T383563#11434840.

rLegal needs to be adjusted to allow for buckets that are fractions,
for example `0.1`, and those that are negative, for example `-1`. This
is covered by the histogram test which contains such buckets.

The regular expression for timings was missing a `$` terminal anchor,
for simplicities sake it is being added here as well.

Bug: T383563
Depends-On: Ie7858b11ad876f25c3799a03eb2b3b9bcc78aedb
Change-Id: Ifbb0a831eeca4a8181d4347b88513f25430f7469