Delay StickyHeaders section click instrumentation for slow loads
Slow loads seemed to be causing the click instrumentation to fire
early, causing the automatic section openings to be recorded as
"unfold" events by the user.

Delay the setup until the `mobile.init` mobile is loaded which is
when StickyHeaders starts its setup. The 300ms timeout wasn't enough
for these cases.

Note that very slow loads (eg 3G or very bad 4G) will result in
the collapsed sections being visible for seconds or even a minute
on large pages before all the sections open. Collapsed sections
can be opened but not re-closed during this portion of time.

Bug: T412857
Change-Id: I88e641e03c0624fbe57d4f1285512e36a7137444